requirements.txt
selenium==3.13.0
requests==2.19.1
Pillow==5.2.0
